refactor : eventManager를 통해 이벤트 관리
