+++ src/basic/main.basic.js
@@ -1,24 +1,5 @@
-export const state = {
-  productList: [
-    { id: 'p1', name: '상품1', val: 10000, q: 50 },
-    { id: 'p2', name: '상품2', val: 20000, q: 30 },
-    { id: 'p3', name: '상품3', val: 30000, q: 20 },
-    { id: 'p4', name: '상품4', val: 15000, q: 0 },
-    { id: 'p5', name: '상품5', val: 25000, q: 10 },
-  ],
-  lastSel: null,
-  bonusPts: 0,
-  totalAmt: 0,
-  itemCnt: 0,
-};
-
-export const elements = {
-  sel: null,
-  addBtn: null,
-  cartDisp: null,
-  sum: null,
-  stockInfo: null,
-};
+import { attachEventListener } from './eventManager.js';
+import { state, elements } from './state.js';
 
 function main() {
   state.productList = [
@@ -100,7 +81,7 @@ function updateSelOpts() {
     elements.sel.appendChild(opt);
   });
 }
-function calcCart() {
+export function calcCart() {
   state.totalAmt = 0;
   state.itemCnt = 0;
   var cartItems = elements.cartDisp.children;
@@ -183,89 +164,4 @@ function updateStockInfo() {
   elements.stockInfo.textContent = infoMsg;
 }
 main();
-elements.addBtn.addEventListener('click', function () {
-  var selItem = elements.sel.value;
-  var itemToAdd = state.productList.find(function (p) {
-    return p.id === selItem;
-  });
-  if (itemToAdd && itemToAdd.q > 0) {
-    var item = document.getElementById(itemToAdd.id);
-    if (item) {
-      var newQty =
-        parseInt(item.querySelector('span').textContent.split('x ')[1]) + 1;
-      if (newQty <= itemToAdd.q) {
-        item.querySelector('span').textContent =
-          itemToAdd.name + ' - ' + itemToAdd.val + '원 x ' + newQty;
-        itemToAdd.q--;
-      } else {
-        alert('재고가 부족합니다.');
-      }
-    } else {
-      var newItem = document.createElement('div');
-      newItem.id = itemToAdd.id;
-      newItem.className = 'flex justify-between items-center mb-2';
-      newItem.innerHTML =
-        '<span>' +
-        itemToAdd.name +
-        ' - ' +
-        itemToAdd.val +
-        '원 x 1</span><div>' +
-        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-        itemToAdd.id +
-        '" data-change="-1">-</button>' +
-        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-        itemToAdd.id +
-        '" data-change="1">+</button>' +
-        '<button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="' +
-        itemToAdd.id +
-        '">삭제</button></div>';
-      elements.cartDisp.appendChild(newItem);
-      itemToAdd.q--;
-    }
-    calcCart();
-    state.lastSel = selItem;
-  }
-});
-elements.cartDisp.addEventListener('click', function (event) {
-  var tgt = event.target;
-  if (
-    tgt.classList.contains('quantity-change') ||
-    tgt.classList.contains('remove-item')
-  ) {
-    var prodId = tgt.dataset.productId;
-    var itemElem = document.getElementById(prodId);
-    var prod = state.productList.find(function (p) {
-      return p.id === prodId;
-    });
-    if (tgt.classList.contains('quantity-change')) {
-      var qtyChange = parseInt(tgt.dataset.change);
-      var newQty =
-        parseInt(itemElem.querySelector('span').textContent.split('x ')[1]) +
-        qtyChange;
-      if (
-        newQty > 0 &&
-        newQty <=
-          prod.q +
-            parseInt(itemElem.querySelector('span').textContent.split('x ')[1])
-      ) {
-        itemElem.querySelector('span').textContent =
-          itemElem.querySelector('span').textContent.split('x ')[0] +
-          'x ' +
-          newQty;
-        prod.q -= qtyChange;
-      } else if (newQty <= 0) {
-        itemElem.remove();
-        prod.q -= qtyChange;
-      } else {
-        alert('재고가 부족합니다.');
-      }
-    } else if (tgt.classList.contains('remove-item')) {
-      var remQty = parseInt(
-        itemElem.querySelector('span').textContent.split('x ')[1]
-      );
-      prod.q += remQty;
-      itemElem.remove();
-    }
-    calcCart();
-  }
-});
+attachEventListener();
